CI: validation gates on every PR — typecheck, build, splice gate, sync rig
Add .github/workflows/ci.yml running four gates on every PR and push to
main: tsc -b, npm run build:single, the splice conformance gate, and the
CRDT convergence rig at SEEDS=300 (deeper than the local default; still
under a second).

The conformance gate moves out of release.mjs into scripts/shell-gate.mjs
— one implementation, two callers (release before signing, CI on every PR
build), so the checks can never drift apart. Callable standalone:
node scripts/shell-gate.mjs <shell.html>.

CI validates and never releases: no secrets, no signing, no publishing —
releases stay local per docs/RELEASING.md, and the workflow header says so
for anyone tempted to add release steps later.
