+++ scripts/release.mjs
@@ -25,6 +25,7 @@ import { cpSync, existsSync, mkdirSync, readFileSync, rmSync, writeFileSync } fr
 import { join, dirname } from 'node:path'
 import { fileURLToPath } from 'node:url'
 import { spliceDoc } from './guestbook-deck.mjs'
+import { gateShell } from './shell-gate.mjs'
 
 const root = dirname(dirname(fileURLToPath(import.meta.url)))
 const args = process.argv.slice(2)
@@ -51,24 +52,9 @@ cpSync(shellSrc, join(site, 'releases/slides/Bento_Slides.bento.html'))
 cpSync(shellSrc, join(site, 'slides/index.html'))
 
 // CONFORMANCE GATE — old updaters are frozen code; every release must
-// satisfy the splice contract they rely on (see postbuild-compress.mjs):
-// plaintext #bento-doc, regex-extractable, balanced script tags, and a
-// v0.1.0-style text splice must produce a well-formed document.
-{
-  const shell = readFileSync(join(site, 'releases/slides/Bento_Slides.bento.html'), 'utf8')
-  const blockRe = /<script type="application\/bento\+json" id="bento-doc">[\s\S]*?<\/script>/
-  if (!blockRe.test(shell)) throw new Error('GATE: #bento-doc block not found/extractable')
-  const opens = (shell.match(/<script[\s>]/g) ?? []).length
-  const closes = shell.split('</scr' + 'ipt>').length - 1
-  if (opens !== closes) throw new Error(`GATE: script tag imbalance (${opens}/${closes})`)
-  const fakeDoc = JSON.stringify({ format: 'bento/slides', probe: '<tag> & </close>' }).replace(/</g, '\\u003c')
-  const spliced = shell.replace(blockRe, `<script type="application/bento+json" id="bento-doc">\n${fakeDoc}\n</scr` + 'ipt>')
-  if (!spliced.includes(fakeDoc)) throw new Error('GATE: splice failed')
-  const opens2 = (spliced.match(/<script[\s>]/g) ?? []).length
-  const closes2 = spliced.split('</scr' + 'ipt>').length - 1
-  if (opens2 !== closes2) throw new Error('GATE: spliced document imbalanced')
-  console.log('conformance gate: old-updater splice contract OK')
-}
+// satisfy the splice contract they rely on. The gate itself lives in
+// scripts/shell-gate.mjs (shared with CI, which runs it on every PR build).
+gateShell(join(site, 'releases/slides/Bento_Slides.bento.html'))
 
 // Sign the manifest against the staged bytes.
 const signArgs = [
